Fix rendering image for single service
In a case where only one shipping service is enabled,
woo might not render it inside a standard shipping method template
but directly in the parent placeholder. In such cases method label
does not exist so we need to inject an image directly into the
parent placeholder.

Reformatted the code and removed the radio button from
location picker modal.

Issue: CS-487